Add SSL support. Patch contributed by Sean Finney. ------------------------------------------------------------------- The new pool_ssl.c provides the necessary functions for negotiating SSL based connections for both frontend and backend connections.  child.c has been updated to request the negotiation in both cases, and pool_stream.c has been modified to use SSL based i/o functions if a connection has successfully negotiated a connection.
SSL related failures should be handled as gracefully as possible.  For
frontend connections, the client is allowed to continue in cleartext
in the case that SSL negotiation was not successful.  For backend
connections, plaintext connections should continue as well, modulo
any pg_hba restrictions on the remote server.  It is also possible to
have connections where only one of the frontend/backend uses SSL.

By default this functionality is off (further testing should be done
and perhaps better controls on certificate verification, etc).

New fields are added to the global pool configuration to allow specifying
SSL related settings:

 * "ssl", global on/off switch (default: off)
 * "ssl_cert", path to SSL public certificate (default: "")
 * "ssl_key", path to SSL private key (default: "")

The sample pgpool.conf files have been updated with comments and sample
values for these new settings.

The autoconf build system has been updated with a new "--with-openssl" option
to allow enabling/disabling SSL support at build time.  For simplicity and
clarity no effort is made to provide extra flexibility similar to the existing
--with-pgsql{,-include-dir,lib-dir} ./configure flags.  In the meantime it's
expected that the necessary flags are provided via CFLAGS/LDFLAGS etc when
calling ./configure.